requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
langchain==0.1.13
transformers==4.38.2
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.6.4
websockets>=10.0
python-multipart>=0.0.5